requirements.txt
fastapi
uvicorn

SQLAlchemy==1.3.22
SQLAlchemy-Utils==0.41.1
alembic==1.5.2
psycopg2==2.8.6
psycopg2-binary==2.9.3 # PythonからPostgreSQLに接続するためのパッケージ
pydantic[email]==1.6.1
python-jose[cryptography]==3.2.0
python-multipart==0.0.6
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4  # パスワードハッシュ化用
email-validator==1.1.3  # メールアドレス検証用